Rename Public API endpoints
