+++ front/pages/api/v1/w/[wId]/data_sources/[dsId]/documents/[documentId]/parents.ts
@@ -1,4 +1,4 @@
-import handler from "@app/pages/api/v1/w/[wId]/vaults/[vId]/data_sources/[dsId]/documents/[documentId]/parents";
+import handler from "@app/pages/api/v1/w/[wId]/spaces/[spaceId]/data_sources/[dsId]/documents/[documentId]/parents";
 
 /**
  * @ignoreswagger
