+++ front/pages/api/v1/w/[wId]/data_sources/[dsId]/tables/csv.ts
@@ -1,4 +1,4 @@
-import handler from "@app/pages/api/v1/w/[wId]/vaults/[vId]/data_sources/[dsId]/tables/csv";
+import handler from "@app/pages/api/v1/w/[wId]/spaces/[spaceId]/data_sources/[dsId]/tables/csv";
 
 export const config = {
   api: {
