+++ front/pages/api/v1/w/[wId]/spaces/[spaceId]/apps/[aId]/runs/[runId]/index.ts
@@ -1,0 +1,156 @@
+import type { RunType, WithAPIErrorResponse } from "@dust-tt/types";
+import { CoreAPI } from "@dust-tt/types";
+import type { NextApiRequest, NextApiResponse } from "next";
+
+import apiConfig from "@app/lib/api/config";
+import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
+import type { Authenticator } from "@app/lib/auth";
+import { AppResource } from "@app/lib/resources/app_resource";
+import { SpaceResource } from "@app/lib/resources/space_resource";
+import logger from "@app/logger/logger";
+import { apiError } from "@app/logger/withlogging";
+
+export const config = {
+  api: {
+    responseLimit: "8mb",
+  },
+};
+
+export type GetRunResponseBody = {
+  run: RunType;
+};
+
+/**
+ * @swagger
+ * /api/v1/w/{wId}/spaces/{spaceId}/apps/{aId}/runs/{runId}:
+ *   get:
+ *     summary: Get an app run
+ *     description: Retrieve a run for an app in the space identified by {spaceId}.
+ *     tags:
+ *       - Apps
+ *     security:
+ *       - BearerAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: wId
+ *         required: true
+ *         description: Unique string identifier for the workspace
+ *         schema:
+ *           type: string
+ *       - in: path
+ *         name: spaceId
+ *         required: true
+ *         description: ID of the space
+ *         schema:
+ *           type: string
+ *       - in: path
+ *         name: aId
+ *         required: true
+ *         description: ID of the app
+ *         schema:
+ *           type: string
+ *       - in: path
+ *         name: runId
+ *         required: true
+ *         description: ID of the run
+ *         schema:
+ *           type: string
+ *     responses:
+ *       200:
+ *         description: The run
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 run:
+ *                   $ref: '#/components/schemas/Run'
+ *       400:
+ *         description: Bad Request. Missing or invalid parameters.
+ *       401:
+ *         description: Unauthorized. Invalid or missing authentication token.
+ */
+
+async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse<WithAPIErrorResponse<GetRunResponseBody>>,
+  auth: Authenticator
+): Promise<void> {
+  const owner = auth.getNonNullableWorkspace();
+
+  // Handling the case where `spaceId` is undefined to keep support for the legacy endpoint (not under
+  // space, global space assumed for the auth (the authenticator associated with the app, not the
+  // user)).
+  let { spaceId } = req.query;
+  if (spaceId === undefined) {
+    spaceId = (await SpaceResource.fetchWorkspaceGlobalSpace(auth)).sId;
+  }
+
+  const app = await AppResource.fetchById(auth, req.query.aId as string);
+
+  if (!app || !app.canRead(auth) || app.space.sId !== spaceId) {
+    return apiError(req, res, {
+      status_code: 404,
+      api_error: {
+        type: "app_not_found",
+        message: "The app you're trying to access was not found",
+      },
+    });
+  }
+
+  switch (req.method) {
+    case "GET":
+      const runId = req.query.runId as string;
+
+      logger.info(
+        {
+          workspace: {
+            sId: owner.sId,
+            name: owner.name,
+          },
+          app: app.sId,
+          runId,
+        },
+        "App run retrieve"
+      );
+
+      const coreAPI = new CoreAPI(apiConfig.getCoreAPIConfig(), logger);
+      const runRes = await coreAPI.getRun({
+        projectId: app.dustAPIProjectId,
+        runId,
+      });
+      if (runRes.isErr()) {
+        return apiError(req, res, {
+          status_code: 400,
+          api_error: {
+            type: "run_error",
+            message: "There was an error retrieving the run.",
+            run_error: runRes.error,
+          },
+        });
+      }
+      const run: RunType = runRes.value.run;
+      run.specification_hash = run.app_hash;
+      delete run.app_hash;
+
+      if (run.status.run === "succeeded" && run.traces.length > 0) {
+        run.results = run.traces[run.traces.length - 1][1];
+      } else {
+        run.results = null;
+      }
+
+      res.status(200).json({ run });
+      return;
+
+    default:
+      return apiError(req, res, {
+        status_code: 405,
+        api_error: {
+          type: "method_not_supported_error",
+          message: "The method passed is not supported, GET is expected.",
+        },
+      });
+  }
+}
+
+export default withPublicAPIAuthentication(handler);
