+++ front/pages/api/v1/w/[wId]/spaces/[spaceId]/data_source_views/[dsvId]/search.ts
@@ -1,0 +1,193 @@
+import type { WithAPIErrorResponse } from "@dust-tt/types";
+import type { NextApiRequest, NextApiResponse } from "next";
+
+import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
+import type { Authenticator } from "@app/lib/auth";
+import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
+import { apiError } from "@app/logger/withlogging";
+import type { DatasourceSearchResponseBody } from "@app/pages/api/w/[wId]/data_sources/[dsId]/search";
+import { handleDataSourceSearch } from "@app/pages/api/w/[wId]/data_sources/[dsId]/search";
+
+/**
+ * @swagger
+ * /api/v1/w/{wId}/spaces/{spaceId}/data_source_views/{dsvId}/search:
+ *   get:
+ *     summary: Search the data source view
+ *     description: Search the data source view identified by {dsvId} in the workspace identified by {wId}.
+ *     tags:
+ *       - DatasourceViews
+ *     security:
+ *       - BearerAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: wId
+ *         required: true
+ *         description: ID of the workspace
+ *         schema:
+ *           type: string
+ *       - in: path
+ *         name: spaceId
+ *         required: true
+ *         description: ID of the space
+ *         schema:
+ *           type: string
+ *       - in: path
+ *         name: dsvId
+ *         required: true
+ *         description: ID of the data source view
+ *         schema:
+ *           type: string
+ *       - in: query
+ *         name: query
+ *         required: true
+ *         description: The search query
+ *         schema:
+ *           type: string
+ *       - in: query
+ *         name: top_k
+ *         required: true
+ *         description: The number of results to return
+ *         schema:
+ *           type: number
+ *       - in: query
+ *         name: full_text
+ *         required: true
+ *         description: Whether to return the full document content
+ *         schema:
+ *           type: boolean
+ *       - in: query
+ *         name: target_document_tokens
+ *         required: false
+ *         description: The number of tokens in the target document
+ *         schema:
+ *           type: number
+ *       - in: query
+ *         name: timestamp_gt
+ *         required: false
+ *         description: The timestamp to filter by
+ *         schema:
+ *           type: number
+ *       - in: query
+ *         name: timestamp_lt
+ *         required: false
+ *         description: The timestamp to filter by
+ *         schema:
+ *           type: number
+ *       - in: query
+ *         name: tags_in
+ *         required: false
+ *         description: The tags to filter by
+ *         schema:
+ *           type: string
+ *       - in: query
+ *         name: tags_not
+ *         required: false
+ *         description: The tags to filter by
+ *         schema:
+ *           type: string
+ *       - in: query
+ *         name: parents_in
+ *         required: false
+ *         description: The parents to filter by
+ *         schema:
+ *           type: string
+ *       - in: query
+ *         name: parents_not
+ *         required: false
+ *         description: The parents to filter by
+ *         schema:
+ *           type: string
+ *     responses:
+ *       200:
+ *         description: The documents
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 documents:
+ *                   type: array
+ *                   items:
+ *                     type: object
+ *                     properties:
+ *                       id:
+ *                         type: string
+ *                         description: ID of the document
+ *                       title:
+ *                         type: string
+ *                         description: Title of the document
+ *                       content:
+ *                         type: string
+ *                         description: Content of the document
+ *                       tags:
+ *                         type: array
+ *                         items:
+ *                           type: string
+ *                         description: Tags of the document
+ *                       parents:
+ *                         type: array
+ *                         items:
+ *                           type: string
+ *                         description: Parents of the document
+ *                       timestamp:
+ *                         type: number
+ *                         description: Timestamp of the document
+ *                       data:
+ *                         type: object
+ *                         description: Data of the document
+ *                       score:
+ *                         type: number
+ *                         description: Score of the document
+ *       400:
+ *         description: Invalid request error
+ *       405:
+ *         description: Method not supported error
+ */
+async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse<WithAPIErrorResponse<DatasourceSearchResponseBody>>,
+  auth: Authenticator
+): Promise<void> {
+  const { dsvId, spaceId } = req.query;
+  if (typeof dsvId !== "string" || typeof spaceId !== "string") {
+    return apiError(req, res, {
+      status_code: 400,
+      api_error: {
+        type: "invalid_request_error",
+        message: "Invalid path parameters.",
+      },
+    });
+  }
+
+  const dataSourceView = await DataSourceViewResource.fetchById(auth, dsvId);
+  if (!dataSourceView || dataSourceView.space.sId !== spaceId) {
+    return apiError(req, res, {
+      status_code: 404,
+      api_error: {
+        type: "data_source_not_found",
+        message: "The data source you requested was not found.",
+      },
+    });
+  }
+
+  switch (req.method) {
+    case "GET": {
+      return handleDataSourceSearch({
+        req,
+        res,
+        dataSource: dataSourceView.dataSource,
+      });
+    }
+
+    default:
+      return apiError(req, res, {
+        status_code: 405,
+        api_error: {
+          type: "method_not_supported_error",
+          message: "The method passed is not supported, GET is expected.",
+        },
+      });
+  }
+}
+
+export default withPublicAPIAuthentication(handler);
