--- front/pages/api/v1/w/[wId]/vaults/[vId]/data_sources/[dsId]/search.ts
@@ -1,211 +1,0 @@
-import type { WithAPIErrorResponse } from "@dust-tt/types";
-import type { NextApiRequest, NextApiResponse } from "next";
-
-import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
-import type { Authenticator } from "@app/lib/auth";
-import { DataSourceResource } from "@app/lib/resources/data_source_resource";
-import { SpaceResource } from "@app/lib/resources/space_resource";
-import { apiError } from "@app/logger/withlogging";
-import type { DatasourceSearchResponseBody } from "@app/pages/api/w/[wId]/data_sources/[dsId]/search";
-import { handleDataSourceSearch } from "@app/pages/api/w/[wId]/data_sources/[dsId]/search";
-
-/**
- * @swagger
- * /api/v1/w/{wId}/vaults/{vId}/data_sources/{dsId}/search:
- *   get:
- *     summary: Search the data source
- *     description: Search the data source identified by {dsId} in the workspace identified by {wId}.
- *     tags:
- *       - Datasources
- *     security:
- *       - BearerAuth: []
- *     parameters:
- *       - in: path
- *         name: wId
- *         required: true
- *         description: ID of the workspace
- *         schema:
- *           type: string
- *       - in: path
- *         name: vId
- *         required: true
- *         description: ID of the vault
- *         schema:
- *           type: string
- *       - in: path
- *         name: dsId
- *         required: true
- *         description: ID of the data source
- *         schema:
- *           type: string
- *       - in: query
- *         name: query
- *         required: true
- *         description: The search query
- *         schema:
- *           type: string
- *       - in: query
- *         name: top_k
- *         required: true
- *         description: The number of results to return
- *         schema:
- *           type: number
- *       - in: query
- *         name: full_text
- *         required: true
- *         description: Whether to return the full document content
- *         schema:
- *           type: boolean
- *       - in: query
- *         name: target_document_tokens
- *         required: false
- *         description: The number of tokens in the target document
- *         schema:
- *           type: number
- *       - in: query
- *         name: timestamp_gt
- *         required: false
- *         description: The timestamp to filter by
- *         schema:
- *           type: number
- *       - in: query
- *         name: timestamp_lt
- *         required: false
- *         description: The timestamp to filter by
- *         schema:
- *           type: number
- *       - in: query
- *         name: tags_in
- *         required: false
- *         description: The tags to filter by
- *         schema:
- *           type: string
- *       - in: query
- *         name: tags_not
- *         required: false
- *         description: The tags to filter by
- *         schema:
- *           type: string
- *       - in: query
- *         name: parents_in
- *         required: false
- *         description: The parents to filter by
- *         schema:
- *           type: string
- *       - in: query
- *         name: parents_not
- *         required: false
- *         description: The parents to filter by
- *         schema:
- *           type: string
- *     responses:
- *       200:
- *         description: The documents
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 documents:
- *                   type: array
- *                   items:
- *                     type: object
- *                     properties:
- *                       id:
- *                         type: string
- *                         description: ID of the document
- *                       title:
- *                         type: string
- *                         description: Title of the document
- *                       content:
- *                         type: string
- *                         description: Content of the document
- *                       tags:
- *                         type: array
- *                         items:
- *                           type: string
- *                         description: Tags of the document
- *                       parents:
- *                         type: array
- *                         items:
- *                           type: string
- *                         description: Parents of the document
- *                       timestamp:
- *                         type: number
- *                         description: Timestamp of the document
- *                       data:
- *                         type: object
- *                         description: Data of the document
- *                       score:
- *                         type: number
- *                         description: Score of the document
- *       400:
- *         description: Invalid request error
- *       405:
- *         description: Method not supported error
- */
-async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<DatasourceSearchResponseBody>>,
-  auth: Authenticator
-): Promise<void> {
-  const { dsId } = req.query;
-  if (typeof dsId !== "string") {
-    return apiError(req, res, {
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "Invalid path parameters.",
-      },
-    });
-  }
-
-  const dataSource = await DataSourceResource.fetchByNameOrId(
-    auth,
-    dsId,
-    // TODO(DATASOURCE_SID): Clean-up
-    { origin: "v1_data_sources_search" }
-  );
-
-  // Handling the case where vId is undefined to keep support for the legacy endpoint (not under
-  // vault, global vault assumed for the auth (the authenticator associated with the app, not the
-  // user)).
-  let { vId } = req.query;
-  if (typeof vId !== "string") {
-    if (auth.isSystemKey()) {
-      // We also handle the legacy usage of connectors that taps into connected data sources which
-      // are not in the global space. If this is a system key we trust it and set the vId to the
-      // dataSource.space.sId.
-      vId = dataSource?.space.sId;
-    } else {
-      vId = (await SpaceResource.fetchWorkspaceGlobalSpace(auth)).sId;
-    }
-  }
-
-  if (!dataSource || dataSource.space.sId !== vId) {
-    return apiError(req, res, {
-      status_code: 404,
-      api_error: {
-        type: "data_source_not_found",
-        message: "The data source you requested was not found.",
-      },
-    });
-  }
-
-  switch (req.method) {
-    case "GET": {
-      return handleDataSourceSearch({ req, res, dataSource });
-    }
-
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
-        },
-      });
-  }
-}
-
-export default withPublicAPIAuthentication(handler);
