--- front/pages/api/v1/w/[wId]/vaults/index.ts
@@ -1,83 +1,0 @@
-/* eslint jsdoc/no-missing-syntax: 0 */ //
-// Disabling jsdoc rule, as we're not yet documentating dust apps endpoints under vaults.
-// We still document the legacy endpoint, which does the same thing.
-// Note: for now, an API key only has access to the global vault.
-import type { SpaceType, WithAPIErrorResponse } from "@dust-tt/types";
-import type { NextApiRequest, NextApiResponse } from "next";
-
-import { withPublicAPIAuthentication } from "@app/lib/api/wrappers";
-import type { Authenticator } from "@app/lib/auth";
-import { SpaceResource } from "@app/lib/resources/space_resource";
-import { apiError } from "@app/logger/withlogging";
-
-export type GetVaultsResponseBody = {
-  vaults: SpaceType[];
-};
-
-/**
- * @swagger
- * /api/v1/w/{wId}/vaults:
- *   get:
- *     summary: List Workspace Vaults
- *     description: Retrieves a list of vaults for the authenticated workspace.
- *     tags:
- *       - Vaults
- *     security:
- *       - BearerAuth: []
- *     parameters:
- *       - in: path
- *         name: wId
- *         required: true
- *         description: Unique string identifier for the workspace
- *         schema:
- *           type: string
- *     responses:
- *       200:
- *         description: Vaults of the workspace
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 vaults:
- *                   type: array
- *                   items:
- *                     $ref: '#/components/schemas/Vault'
- *       400:
- *         description: Bad Request. Missing or invalid parameters.
- *       401:
- *         description: Unauthorized. Invalid or missing authentication token.
- *       404:
- *         description: Workspace not found.
- *       405:
- *         description: Method not supported.
- *       500:
- *         description: Internal Server Error.
- */
-async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<GetVaultsResponseBody>>,
-  auth: Authenticator
-): Promise<void> {
-  switch (req.method) {
-    case "GET":
-      const vaults = (await SpaceResource.listWorkspaceSpaces(auth)).map(
-        (vault) => vault.toJSON()
-      );
-      res.status(200).json({
-        vaults,
-      });
-      return;
-
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
-        },
-      });
-  }
-}
-
-export default withPublicAPIAuthentication(handler);
